[pslib] pslib with cmake (#32800)
* pslib with cmake

* heter util

* vlog

* heter server test

* add dtor

* cmake
